Controladores y CRUD de las vistas principales

+++ database/migrations/2018_11_20_000007_create_Telefonia_table.php
@@ -25,7 +25,7 @@ class CreateTelefoniaTable extends Migration
             $table->engine = 'InnoDB';
             $table->increments('id');
             $table->string('nombre', 45)->nullable();
-            $table->decimal('precio', 8, 2)->nullable();
+            $table->double('precio')->nullable();
             $table->integer('cantMin')->nullable();
             $table->nullableTimestamps();
         });
